[TAN-8601] Move the hero banner form fields to shared components
