+++ front/app/components/LandingPages/citizen/FixedRatioLayout.ts
@@ -8,7 +8,7 @@ import styled from 'styled-components';
 import {
   homepageBannerLayoutHeights,
   FIXED_RATIO_LAYOUT_ASPECT_RATIO,
-} from 'containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/HeaderImageDropzone';
+} from 'components/admin/GenericHeroBannerForm/HeaderImageDropzone';
 
 export const Container = styled.div`
   width: 100%;
