+++ front/app/components/LandingPages/citizen/FullWidthBannerLayout.ts
@@ -1,7 +1,7 @@
 import { media, colors } from '@citizenlab/cl2-component-library';
 import styled from 'styled-components';
 
-import { homepageBannerLayoutHeights } from 'containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/HeaderImageDropzone';
+import { homepageBannerLayoutHeights } from 'components/admin/GenericHeroBannerForm/HeaderImageDropzone';
 
 export const Container = styled.div`
   width: 100%;
