+++ front/app/components/LandingPages/citizen/TwoRowLayout.ts
@@ -1,8 +1,7 @@
 import { media } from '@citizenlab/cl2-component-library';
 import styled from 'styled-components';
 
-import { homepageBannerLayoutHeights } from 'containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/HeaderImageDropzone';
-
+import { homepageBannerLayoutHeights } from 'components/admin/GenericHeroBannerForm/HeaderImageDropzone';
 import Image from 'components/UI/Image';
 
 export const Container = styled.div`
