+++ front/app/components/admin/GenericHeroBannerForm/BannerImageFields/BannerImageFields.test.tsx
