+++ front/app/components/admin/GenericHeroBannerForm/BannerImageFields/ImageInfoTooltip.tsx
