+++ front/app/components/admin/GenericHeroBannerForm/BannerImageFields/OverlayControls.tsx
