+++ front/app/components/admin/GenericHeroBannerForm/BannerImageFields/index.tsx
