+++ front/app/components/admin/GenericHeroBannerForm/HeaderImageDropzone.test.tsx
