+++ front/app/components/admin/GenericHeroBannerForm/HeaderImageDropzone.tsx
