+++ front/app/components/admin/GenericHeroBannerForm/LayoutSettingField/LayoutSettingField.test.tsx
