+++ front/app/components/admin/GenericHeroBannerForm/LayoutSettingField/layout_previews/two_row_layout_inactive.jpg
