+++ front/app/containers/HomePage/SignedInHeader/index.tsx
@@ -8,8 +8,8 @@ import useCurrentOnboardingCampaign from 'api/onboarding_campaigns/useCurrentOnb
 import useDismissOnboardingCampaign from 'api/onboarding_campaigns/useDismissOnboardingCampaign';
 
 import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/HomepageBanner';
-import { homepageBannerLayoutHeights } from 'containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/HeaderImageDropzone';
 
+import { homepageBannerLayoutHeights } from 'components/admin/GenericHeroBannerForm/HeaderImageDropzone';
 import {
   Container,
   HeaderImage,
