+++ front/app/containers/HomePage/SignedOutHeader/FullWidthBannerLayout.tsx
@@ -4,8 +4,8 @@ import { media } from '@citizenlab/cl2-component-library';
 import styled from 'styled-components';
 
 import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/HomepageBanner';
-import { homepageBannerLayoutHeights } from 'containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/HeaderImageDropzone';
 
+import { homepageBannerLayoutHeights } from 'components/admin/GenericHeroBannerForm/HeaderImageDropzone';
 import {
   Container,
   HeaderImage,
