+++ front/app/containers/HomePage/SignedOutHeader/TwoColumnLayout.tsx
@@ -4,9 +4,9 @@ import { media } from '@citizenlab/cl2-component-library';
 import styled from 'styled-components';
 
 import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/HomepageBanner';
-import { homepageBannerLayoutHeights } from 'containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/HeaderImageDropzone';
 import HeaderContent from 'containers/HomePage/SignedOutHeader/HeaderContent';
 
+import { homepageBannerLayoutHeights } from 'components/admin/GenericHeroBannerForm/HeaderImageDropzone';
 import {
   Container,
   HeaderImageWrapper,
